Fix the test cp issue in LLM based test generation
