+++ src/main/kotlin/org/jetbrains/research/testgenie/tools/ToolUtils.kt
@@ -1,7 +1,10 @@
 package org.jetbrains.research.testgenie.tools
 
 import com.intellij.openapi.components.service
+import com.intellij.openapi.module.ModuleManager
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.roots.CompilerModuleExtension
+import com.intellij.openapi.roots.ModuleRootManager
 import org.jetbrains.research.testgenie.data.Report
 import org.jetbrains.research.testgenie.editor.Workspace
 import org.jetbrains.research.testgenie.tools.evosuite.Pipeline
@@ -59,4 +62,32 @@ fun receiveGenerationResult(
 
 fun cancelPendingResult(project: Project, testResultName: String) {
     project.service<Workspace>().testGenerationData.pendingTestResults.remove(testResultName)
+}
+
+fun getBuildPath(project: Project): String {
+    var buildPath = ""
+
+    for (module in ModuleManager.getInstance(project).modules) {
+        val compilerOutputPath = CompilerModuleExtension.getInstance(module)?.compilerOutputPath
+        compilerOutputPath?.let { buildPath += compilerOutputPath.path.plus(":") }
+
+        // Include extra libraries in classpath
+        val librariesPaths = ModuleRootManager.getInstance(module).orderEntries().librariesOnly().pathsList.pathList
+        for (lib in librariesPaths) {
+            // exclude the invalid classpaths
+            if (buildPath.contains(lib)) {
+                continue
+            }
+            if (lib.endsWith(".zip")) {
+                continue
+            }
+            buildPath += lib.plus(":")
+//            val basePath = module.project.basePath
+//            if (lib.startsWith(basePath.toString())) {
+//                // add the extra path
+
+//            }
+        }
+    }
+    return buildPath
 }
